+++ OzTrack/Crusher.m
@@ -9,7 +9,7 @@
 #import "Crusher.h"
 
 @interface Crusher()
-@property(nonatomic) float tonsAtCrusher;
+@property(nonatomic) CGFloat tonsAtCrusher;
 @property(strong, nonatomic) NSArray *currentGradesAtCrusher;
 @property(strong, nonatomic) NSArray *targetGradesAtCrusher;
 @property(strong, nonatomic) NSArray *dailyVariation;
@@ -20,7 +20,7 @@
 
 @implementation Crusher
 
-const float MAX_TONS_HOUR = 1700.0;
+const CGFloat MAX_TONS_HOUR = 1700.0;
 
 -(BOOL)canDirectTip:(Truck *)truck
 {
@@ -28,7 +28,7 @@ const float MAX_TONS_HOUR = 1700.0;
         return NO;
     }
     for (int i = 0; i < [[self currentGradesAtCrusher]count]; i++) {
-        float weightage = ([self tonsAtCrusher] * [[[self currentGradesAtCrusher]objectAtIndex:i]floatValue] ) + (truck.tonsInTruck * [[[truck currentGradesAtTruck]objectAtIndex:i]floatValue] );
+        CGFloat weightage = ([self tonsAtCrusher] * [[[self currentGradesAtCrusher]objectAtIndex:i]floatValue] ) + (truck.tonsInTruck * [[[truck currentGradesAtTruck]objectAtIndex:i]floatValue] );
         float grade = weightage/([self tonsAtCrusher] + [truck tonsInTruck]);
         if(grade > [[[self targetGradesAtCrusher]objectAtIndex:i]floatValue] +[[[self targetGradesAtCrusher]objectAtIndex:i]floatValue] * [[[self hourlyVariation]objectAtIndex:i]floatValue]/100  || grade < [[[self targetGradesAtCrusher]objectAtIndex:i]floatValue] - [[[self targetGradesAtCrusher]objectAtIndex:i]floatValue] * [[[self hourlyVariation]objectAtIndex:i]floatValue]/100 )
         {
